+++ CMakeLists.txt
@@ -152,14 +152,18 @@ install(FILES ${g2o_PATH} DESTINATION lib)
 add_executable(orbslam3_mono_node
 	src/node.cpp
 	src/slam.cpp
+	src/monocular.cpp
 	src/orbslam3/monocular.cpp
 	src/orbslam3/monocular_driver.cpp
 )
+target_compile_definitions(orbslam3_mono_node PRIVATE ORBSLAM3=1)
 add_executable(morbslam_mono_node
 	src/node.cpp
 	src/slam.cpp
+	src/monocular.cpp
 	src/morbslam/monocular.cpp
 )
+target_compile_definitions(morbslam_mono_node PRIVATE MORBSLAM=1)
 
 ament_target_dependencies(orbslam3_mono_node
   PUBLIC ${THIS_PACKAGE_INCLUDE_DEPENDS}
